New CameraProvider, moved all remote providers to separate files
Added a menu to choose the image provider
